% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{example_question_index}
\alias{example_question_index}
\title{Path to the example question index}
\usage{
example_question_index()
}
\value{
The absolute path to the package's example index CSV.
}
\description{
Path to the example question index
}
\examples{
example_question_index()
}
